Remove dependency on torch_cluster
