+++ torchmdnet/priors/coulomb.py
@@ -1,7 +1,7 @@
 import torch
 from torchmdnet.priors.base import BasePrior
-from torchmdnet.models.utils import Distance
 from torch_scatter import scatter
+from torchmdnet.models.utils import OptimizedDistance
 from typing import Optional, Dict
 
 class Coulomb(BasePrior):
@@ -20,7 +20,7 @@ class Coulomb(BasePrior):
             distance_scale = dataset.distance_scale
         if energy_scale is None:
             energy_scale = dataset.energy_scale
-        self.distance = Distance(0, torch.inf, max_num_neighbors=max_num_neighbors)
+        self.distance = OptimizedDistance(0, torch.inf, max_num_pairs=-max_num_neighbors)
         self.alpha = alpha
         self.max_num_neighbors = max_num_neighbors
         self.distance_scale = float(distance_scale)
